feat: add player money management and saving account retrieval
Implemented AddPlayerMoney and GetPlayerSavingAccount methods in the
player use case and repository, allowing players to add funds and
retrieve their saving account balance. Updated the PlayerHttpHandler to
include new endpoints for these functionalities. Enhanced routing to
support the new features.

+++ modules/player/playerModel.go
@@ -23,7 +23,7 @@ type (
 	}
 
 	CreatePlayerTransactionReq struct {
-		PlayerId string `json:"player_id" validate:"required,max=64"`
-		Amount   int    `json:"amount" validate:"required"`
+		PlayerId string  `json:"player_id" validate:"required,max=64"`
+		Amount   float64 `json:"amount" validate:"required"`
 	}
 )
